requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
jikanpy==4.3.2
multidict==5.2.0
python-decouple==3.5
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.31
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarl==1.7.2
